- replaced the crucial 'int's with 'Py_ssize_t'; int's in type slots are not replaced yet

+++ source/orange/cls_orange.cpp
@@ -463,7 +463,7 @@ PyObject *packOrangeDictionary(PyObject *self)
 int unpackOrangeDictionary(PyObject *self, PyObject *dict)
 {
   PyObject *d_key, *d_value;
-  int i = 0;
+  Py_ssize_t i = 0;
   while (PyDict_Next(dict, &i, &d_key, &d_value)) {
 	  if (Orange_setAttrLow(self, d_key, d_value, false) == -1)
 	    return -1;
